# openapi.yaml
openapi: 3.1.0
info:
  title: Fintech API
  version: 0.0.1
  description: Public contract for authenticated transaction creation with account and category ownership.
servers:
  - url: https://fintech-api-87yw.onrender.com
    description: Production
  - url: http://localhost:3000
    description: Local development
security: []
paths:
  /auth/register:
    post:
      summary: Register with email and password
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '201':
          $ref: '#/components/responses/AuthTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/login:
    post:
      summary: Authenticate with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          $ref: '#/components/responses/AuthTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/apple:
    post:
      summary: Authenticate with Sign in with Apple
      operationId: authenticateWithApple
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppleAuthenticationRequest'
      responses:
        '200':
          $ref: '#/components/responses/AuthTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/AccountLinkRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /auth/refresh:
    post:
      summary: Rotate a refresh token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          $ref: '#/components/responses/AuthTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/logout:
    post:
      summary: Revoke a refresh session
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '204':
          description: Session revoked; no response body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /accounts:
    get:
      summary: List accounts owned by the authenticated user
      operationId: listAccounts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Accounts owned by the authenticated user. An empty list is valid.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create an account for the authenticated user
      operationId: createAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /transactions:
    post:
      summary: Create or replay a transaction
      operationId: createTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: UUID that remains stable for every retry of the same complete request payload.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '201':
          description: Transaction created, or the original transaction returned for an identical replay.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /health:
    get:
      summary: Check process liveness
      description: Confirms only that the application process is responding; it does not check database readiness.
      operationId: getHealth
      responses:
        '200':
          description: Application process is responding.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    RequestId:
      description: Server-generated request correlation UUID.
      schema:
        type: string
        format: uuid
  schemas:
    Credentials:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Normalized by trimming surrounding whitespace and lowercasing.
        password:
          type: string
          minLength: 12
          maxLength: 128
          format: password
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          maxLength: 256
    AppleAuthenticationRequest:
      type: object
      additionalProperties: false
      required: [identityToken, authorizationCode, nonce]
      properties:
        identityToken:
          type: string
          minLength: 1
          maxLength: 16384
          description: Compact RS256 identity token returned by Apple.
        authorizationCode:
          type: string
          minLength: 1
          maxLength: 4096
          description: Single-use authorization code returned by Apple.
        nonce:
          type: string
          pattern: '^[A-Za-z0-9_-]{43}$'
          description: Exact unpadded Base64URL raw nonce sent to Apple for this authorization attempt.
    AuthResponse:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken, tokenType, expiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          const: Bearer
        expiresIn:
          type: integer
          const: 900
    CreateAccountRequest:
      type: object
      additionalProperties: false
      required: [name, currency]
      properties:
        name:
          type: string
          minLength: 1
          description: User-provided account name; leading and trailing whitespace is removed before persistence, and whitespace-only names are invalid.
        currency:
          type: string
          const: BRL
          description: BRL is the only supported account currency in the current MVP.
    AccountResponse:
      type: object
      additionalProperties: false
      required: [id, name, currency, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        currency:
          type: string
          const: BRL
        createdAt:
          type: string
          format: date-time
    TransactionRequest:
      type: object
      additionalProperties: false
      required:
        - accountId
        - type
        - amountMinor
        - currency
        - occurredAt
        - clientMutationId
      properties:
        accountId:
          type: string
          format: uuid
        categoryId:
          type: [string, 'null']
          format: uuid
        type:
          type: string
          enum: [expense, income]
          description: expense is an outgoing amount; income is an incoming amount.
        amountMinor:
          type: integer
          minimum: 1
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        description:
          type: [string, 'null']
        occurredAt:
          type: string
          format: date-time
        clientMutationId:
          type: string
          format: uuid
          description: Stable for retries of one logical operation; not independently persisted or database-unique in this milestone.
    TransactionResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - accountId
        - categoryId
        - type
        - amountMinor
        - currency
        - description
        - occurredAt
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        categoryId:
          type: [string, 'null']
          format: uuid
        type:
          type: string
          enum: [expense, income]
          description: expense is an outgoing amount; income is an incoming amount.
        amountMinor:
          type: integer
          minimum: 1
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        description:
          type: [string, 'null']
        occurredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
    AccountLinkRequiredError:
      type: object
      additionalProperties: false
      required: [statusCode, message, code]
      properties:
        statusCode:
          type: integer
          const: 409
        message:
          type: string
          const: Account linking required
        code:
          type: string
          const: ACCOUNT_LINK_REQUIRED
  responses:
    AuthTokens:
      description: Access token and a new opaque refresh token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthResponse'
    BadRequest:
      description: Invalid header or request body.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Unique resource conflict or idempotency key reused with a different payload.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    AccountLinkRequired:
      description: The verified Apple email belongs to another user; explicit linking is required.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AccountLinkRequiredError'
    Unauthorized:
      description: Missing, invalid, or expired authentication credential.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Account or category does not exist for the authenticated user. Foreign-owned resources are intentionally indistinguishable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Authentication endpoint rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Unexpected server or persistence failure.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Apple authentication is temporarily unavailable or misconfigured.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'